+++ R/utils.r
@@ -76,10 +76,11 @@ try_sort_numeric <- function(x,
     sort(x, decreasing = decreasing, ...)
   }
 }
-try_sort_numeric(letters)
-  coercible <- lighthouse::is_coercible_numeric(letters, na = "TRUE")
 
-class_collapse <- function(x, sep = ", ") stringr::str_c(class(x), collapse = sep)
+class_collapse <- function(x, sep = ", ") {
+  stringr::str_c(class(x), collapse = sep)
+}
+
 strip_html <- function(x) {
   stopifnot(is.character(x))
   has_tags <- grepl("<[A-Za-z!/]", x)
@@ -172,7 +173,6 @@ cb_match_type <- function(nm,
 
 as_named <- function(x, class) setNames(as(x, class), names(x))
 
-
 has_val_labels <- function(x) !is.null(labelled::val_labels(x))
 
 to_labelled_chr <- function(x, 
@@ -185,9 +185,6 @@ to_labelled_chr <- function(x,
   )
 }
 
-spread_if_any <- function(..., na.rm = TRUE) {
-  lighthouse::max_if_any(..., na.rm = na.rm) - lighthouse::min_if_any(..., na.rm = na.rm)
-}
 
 #' @export
 nan_to_na.default <- function(x) dplyr::if_else(is.nan(x), NA, x)